+++ pachca_test.go
@@ -185,6 +185,12 @@ func (s *PachcaSuite) TestNilClient(c *C) {
 	err = cc.UpdateBot(0, "")
 	c.Assert(err, Equals, ErrNilClient)
 
+	_, err = cc.GetWebhookEvents(1)
+	c.Assert(err, Equals, ErrNilClient)
+
+	err = cc.DeleteWebhookEvent("00000000000000000000000000")
+	c.Assert(err, Equals, ErrNilClient)
+
 	// FORMS
 
 	err = cc.OpenView(nil)
@@ -366,6 +372,12 @@ func (s *PachcaSuite) TestErrors(c *C) {
 	c.Assert(err, Equals, ErrInvalidBotID)
 	err = cc.UpdateBot(1, "")
 	c.Assert(err, Equals, ErrEmptyFilePath)
+	_, err = cc.GetWebhookEvents(0)
+	c.Assert(err, Equals, ErrInvalidMaxPages)
+	err = cc.DeleteWebhookEvent("")
+	c.Assert(err, Equals, ErrInvalidEventID)
+	err = cc.DeleteWebhookEvent("0000")
+	c.Assert(err, Equals, ErrInvalidEventID)
 
 	// FORMS